fix(compression): keep quantumlock selector under cap
