+++ open-sse/services/compression/strategySelector.ts
@@ -34,10 +34,12 @@ import {
 import { resolveAdaptivePlan } from "./adaptiveCompression/resolveAdaptivePlan.ts";
 import type { AdaptiveTelemetry } from "./adaptiveCompression/types.ts";
 import type { RiskGateConfig } from "./riskGate/riskGate.ts";
+import { resolveRiskGate, withRiskGateAsync } from "./riskGate/strategyWrap.ts";
 import {
   withCompressionEntrypointGuards,
   withCompressionEntrypointGuardsAsync,
 } from "./entrypointWrap.ts";
+export { resolveCacheAwareConfig } from "./cacheAwareConfig.ts";
 
 // Re-export so existing importers (resolver test + chatCore dynamic import) keep resolving.
 export { planFromHeader, formatCompressionMeta, buildNamedComboLookup };
@@ -223,32 +225,6 @@ export function selectCompressionStrategy(
 ): CompressionMode {
   return selectCompressionPlan(config, comboId, estimatedTokens, body, context, combos, header)
     .mode as CompressionMode;
-}
-
-/**
- * #3890: honor the cache-aware `skipSystemPrompt` decision that `getCacheAwareStrategy`
- * already computes but that `selectCompressionStrategy` (which can only return a mode
- * string) previously discarded. In a caching context the system prompt is part of the
- * cacheable prefix, so compressing it breaks the upstream prompt cache. This forces
- * `preserveSystemPrompt` on for caching requests even when the operator turned it off,
- * and leaves non-caching requests untouched.
- */
-export function resolveCacheAwareConfig(
-  config: CompressionConfig,
-  body?: Record<string, unknown>,
-  context?: CachingDetectionContext
-): CompressionConfig {
-  if (!body) return config;
-  const ctx = detectCachingContext(body, context);
-  // Only `skipSystemPrompt` is consumed here, and it depends solely on `ctx.isCachingProvider`
-  // (NOT on the strategy arg — see getCacheAwareStrategy), so the stored `defaultMode` is a safe
-  // input even though it may be "off" for a panel-configured install. If getCacheAwareStrategy is
-  // ever extended to key `skipSystemPrompt` on the mode, pass the resolved effective mode instead.
-  const cacheAware = getCacheAwareStrategy(config.defaultMode, ctx);
-  if (cacheAware.skipSystemPrompt && config.preserveSystemPrompt === false) {
-    return { ...config, preserveSystemPrompt: true };
-  }
-  return config;
 }
 
 export function applyCompression(
